Code/backend/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
flask>=2.0.0
flask-cors>=3.0.10
opencv-python>=4.5.0
numpy>=1.19.0
pillow>=8.0.0
